Move the buffered reader to it’s own module.
